round 2 code reviews added
